extract event dropdown into components
