+++ src/core/dashboard/calendar/event-menu/DeleteEvent.tsx
@@ -1,0 +1,65 @@
+import { TrashIcon } from 'lucide-react';
+import { useState } from 'react';
+import { Button } from '~/components/ui/Button';
+import {
+  Dialog,
+  DialogContent,
+  DialogDescription,
+  DialogFooter,
+  DialogHeader,
+  DialogTitle,
+  DialogTrigger,
+} from '~/components/ui/Dialog';
+import { DropdownMenuItem } from '~/components/ui/DropdownMenu';
+import { api, type RouterOutputs } from '~/utils/api';
+
+export default function DeleteEventDialog({
+  event: { id: eventId },
+}: {
+  event: RouterOutputs['event']['fromTo'][number]['event'];
+}) {
+  const [dialogOpen, setDialogOpen] = useState(false);
+
+  const utils = api.useContext();
+  const { mutateAsync: deleteEvent, isLoading: isDeleteLoading } =
+    api.event.delete.useMutation({
+      onSuccess: async () => {
+        await utils.event.invalidate();
+        setDialogOpen(false);
+      },
+    });
+
+  return (
+    <Dialog open={dialogOpen} onOpenChange={setDialogOpen}>
+      <DialogTrigger asChild>
+        <DropdownMenuItem
+          onSelect={e => e.preventDefault()}
+          className='text-red-500 focus:text-red-500'
+        >
+          <TrashIcon className='mr-2 h-4 w-4' />
+          <span>Delete event</span>
+        </DropdownMenuItem>
+      </DialogTrigger>
+      <DialogContent>
+        <DialogHeader>
+          <DialogTitle>Are you sure absolutely sure?</DialogTitle>
+          <DialogDescription>
+            This action cannot be undone. Are you sure you want to permanently
+            delete this event?
+          </DialogDescription>
+        </DialogHeader>
+        <DialogFooter>
+          <Button
+            variant='destructive'
+            loading={isDeleteLoading}
+            onClick={async () => {
+              await deleteEvent({ id: eventId });
+            }}
+          >
+            Delete event
+          </Button>
+        </DialogFooter>
+      </DialogContent>
+    </Dialog>
+  );
+}
